move RunResult and RunSettings to bridge
